src/lab4/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
pycryptodome=3.15.0=py310h695f6f0_0
